requirements.txt
apscheduler==3.7.0
requests==2.25.0
python-decouple==3.3
prometheus-client==0.9.0
pygithub==1.51